easytools

+++ install.sh
@@ -14,5 +14,8 @@ ln -s "/easytools/menu.sh" "/usr/local/bin/et"
 # Run 'et' command (make sure 'et' is set up before this point)
 et
 
-#
-rm -rf /root/install.sh
+# Check if it's not being run for removal
+if [[ "${BASH_SOURCE[0]}" == "${0}" ]]; then
+    rm -rf /root/install.sh
+fi
+
